+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 - Inbox dao event
 - Inbox proposal event
+- Add natsclient publisher
 
 ## [0.0.7] - 2023-06-07